separate md fixtures
